Move logo function to logo.py
