+++ pygmt/src/__init__.py
@@ -2,4 +2,8 @@
 Source code for PyGMT modules.
 """
 # pylint: disable=import-outside-toplevel
+# GMT Primary modules
+from pygmt.src.logo import logo
+
+# GMT Supplementary modules
 from pygmt.src.meca import meca
